# Go_Web/Clase4T/docs/swagger.yaml
definitions:
  handler.request:
    properties:
      age:
        type: integer
      creation_date:
        type: string
      email:
        type: string
      height:
        type: integer
      last_name:
        type: string
      name:
        type: string
      status:
        type: boolean
    required:
    - age
    - creation_date
    - email
    - height
    - last_name
    - name
    - status
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.mx/support
  description: This API Handle MELI Users
  title: MELI BOOTCAMP API
  version: "1.0"
paths:
  /users:
    get:
      description: store users
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List Users
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found Users
          schema:
            $ref: '#/definitions/web.Response'
      summary: List Users
      tags:
      - Users
    post:
      description: get users
      parameters:
      - description: token requerido
        in: header
        name: token
        required: true
        type: string
      - description: User to Store
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found Users
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store Users
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: Id User
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found Users
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      parameters:
      - description: Id User
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        type: string
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found Users
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update Name-LastName users
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: Id User
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        type: string
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not found Users
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update users
      tags:
      - Users
swagger: "2.0"